Add multy users

+++ my-anime-app/src/api/FavoriteWatchlist.ts
@@ -3,8 +3,10 @@ import type { BaseAnime } from '../types/anime.types.ts';
 
 export const Favorites = {
   async add(anime: BaseAnime) {
+    const { data: { user } } = await supabase.auth.getUser()
     const { error } = await supabase.from('favorites').insert([
       {
+        user_id: user?.id,
         mal_id: anime.mal_id,
         title: anime.title,
         episodes: anime.episodes,
@@ -17,19 +19,23 @@ export const Favorites = {
   },
 
   async remove(mal_id: number) {
+    const { data: { user } } = await supabase.auth.getUser()
     const { error } = await supabase
       .from('favorites')
       .delete()
-      .eq('mal_id', mal_id);
+      .eq('mal_id', mal_id)
+      .eq('user_id', user?.id);
 
     return !error;
   },
 
   async isAdded(mal_id: number): Promise<boolean> {
+    const { data: { user } } = await supabase.auth.getUser()
     const { data } = await supabase
       .from('favorites')
       .select('mal_id')
       .eq('mal_id', mal_id)
+      .eq('user_id', user?.id)
       .single();
 
     return !!data;
@@ -38,8 +44,10 @@ export const Favorites = {
 
 export const Watchlist = {
   async add(anime: BaseAnime) {
+     const { data: { user } } = await supabase.auth.getUser()
     const { error } = await supabase.from('watchlist').insert([
       {
+         user_id: user?.id,
         mal_id: anime.mal_id,
         title: anime.title,
         episodes: anime.episodes,
@@ -56,19 +64,23 @@ export const Watchlist = {
   },
 
   async remove(mal_id: number) {
+     const { data: { user } } = await supabase.auth.getUser()
     const { error } = await supabase
       .from('watchlist')
       .delete()
-      .eq('mal_id', mal_id);
+      .eq('mal_id', mal_id)
+      .eq('user_id', user?.id);
 
     return !error;
   },
 
   async isAdded(mal_id: number): Promise<boolean> {
+     const { data: { user } } = await supabase.auth.getUser()
     const { data } = await supabase
       .from('watchlist')
       .select('mal_id')
       .eq('mal_id', mal_id)
+      .eq('user_id', user?.id)
       .maybeSingle();
 
     return !!data;
